Add device relations and enhanced stock reservations
Introduce device composition relations with audit logging and expand stock reservation capabilities.

Key changes:
- Device relations: add DeviceRelation, DeviceRelationAuditLog, DeviceRelationType, audit action enum, DTOs, view, EF configurations, repository methods, service (create/list/delete with cycle detection), and controller endpoints; register DbSets and DI for DeviceRelationService.
- Stock binding enhancements: add ReservedAmount and OwnerProcessGuid to StockBinding, expose in DTO/view/configuration, validate reserved amounts, compute overlapping reserved quantity, support owner-scoped bindings and deletions, batch assignment API accepting quantities, and availability checks by amount.
- Repository: new queries for device relations, audit logs, owner-scoped bindings, and overlapping reserved amount; include relation navigation properties in device loading.
- Rentals: item quantity types changed to long; handlers updated to use the new assignment API, set owner process GUID on bindings, filter checklist by owner process, and restrict removals to the owning process.
- Misc: ModelBuilder extension now accepts both "Persistence" and misspelled "Persistance" namespaces; minor formatting fix in DeviceController.

These changes enable hierarchical device composition management with audit trails and more robust reservation semantics for rentals (per-process scoping and quantity checks).

+++ LumenForgeServer/LumenForgeServer/Rentals/Dto/Command/ItemActionDtos.cs
@@ -13,7 +13,7 @@ public sealed record ItemAssignmentDto
 
     /// <summary>Number of units to reserve.</summary>
     [JsonPropertyName("quantity")]
-    public required int Quantity { get; init; }
+    public required long Quantity { get; init; }
 }
 
 /// <summary>API request DTO for assigning inventory items to a rental.</summary>
